PUNTO2/requirements.txt
flask
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
Pillow
requests
